refactor: rebuild file layer strcture for home page
