+++ Frontend/WeatherJYJAM_frontend/src/pages/Home/_components/Header/Logo/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './Logo'
